Reimport log.c into source tree
This library is a source only library. Hence those are meant to be
integrated as such in projects (as they are in main branch). It will be
very inconvenient to fetch from another project when building with cmake
when it is be reintroduced, hence the reimport.

They are imported from commit f9ea34994bd58ed342d2245cd4110bb5c6790153
as prescribed in build.zig.zon.